+++ packages/waku/src/lib/plugins/vite-plugin-deploy-deno.ts
@@ -1,37 +1,47 @@
-import path from 'node:path';
-import { existsSync } from 'node:fs';
-import { normalizePath } from 'vite';
 import type { Plugin } from 'vite';
 
 import { unstable_getPlatformObject } from '../../server.js';
-import { EXTENSIONS, SRC_ENTRIES } from '../constants.js';
-import {
-  decodeFilePathFromAbsolute,
-  extname,
-  fileURLToFilePath,
-  joinPath,
-} from '../utils/path.js';
-import { DIST_SERVE_JS, DIST_PUBLIC } from '../builder/constants.js';
+import { SRC_ENTRIES } from '../constants.js';
+import { DIST_PUBLIC } from '../builder/constants.js';
 
-const resolveFileName = (fname: string) => {
-  for (const ext of EXTENSIONS) {
-    const resolvedName = fname.slice(0, -extname(fname).length) + ext;
-    if (existsSync(resolvedName)) {
-      return resolvedName;
-    }
+const SERVE_JS = 'serve-deno.js';
+
+const getServeJsContent = (
+  distDir: string,
+  distPublic: string,
+  srcEntriesFile: string,
+) => `
+import { Hono } from 'https://deno.land/x/hono/mod.ts';
+import { serveStatic } from 'https://deno.land/x/hono/middleware.ts';
+import { runner } from 'waku/unstable_hono';
+
+const distDir = '${distDir}';
+const publicDir = '${distPublic}';
+const loadEntries = () => import('${srcEntriesFile}');
+const env = Deno.env.toObject();
+
+const app = new Hono();
+app.use('*', serveStatic({ root: distDir + '/' + publicDir }));
+app.use('*', runner({ cmd: 'start', loadEntries, env }));
+app.notFound(async (c) => {
+  const file = distDir + '/' + publicDir + '/404.html';
+  const info = await Deno.stat(file);
+  if (info.isFile) {
+    c.header('Content-Type', 'text/html; charset=utf-8');
+    return c.body(await Deno.readFile(file), 404);
   }
-  return fname; // returning the default one
-};
+  return c.text('404 Not Found', 404);
+});
 
-const srcServeFile = decodeFilePathFromAbsolute(
-  joinPath(fileURLToFilePath(import.meta.url), '../../builder/serve-deno.js'),
-);
+Deno.serve(app.fetch);
+`;
 
 export function deployDenoPlugin(opts: {
   srcDir: string;
   distDir: string;
 }): Plugin {
   const platformObject = unstable_getPlatformObject();
+  let entriesFile: string;
   return {
     name: 'deploy-deno-plugin',
     config(viteConfig) {
@@ -39,34 +49,18 @@ export function deployDenoPlugin(opts: {
       if (unstable_phase !== 'buildServerBundle' || deploy !== 'deno') {
         return;
       }
-
-      // FIXME This seems too hacky (The use of viteConfig.root, '.', path.resolve and resolveFileName)
-      const entriesFile = normalizePath(
-        resolveFileName(
-          path.resolve(
-            viteConfig.root || '.',
-            opts.srcDir,
-            SRC_ENTRIES + '.jsx',
-          ),
-        ),
-      );
       const { input } = viteConfig.build?.rollupOptions ?? {};
       if (input && !(typeof input === 'string') && !(input instanceof Array)) {
-        input[DIST_SERVE_JS.replace(/\.js$/, '')] = srcServeFile;
+        input[SERVE_JS.replace(/\.js$/, '')] = `${opts.srcDir}/${SERVE_JS}`;
       }
-      viteConfig.define = {
-        ...viteConfig.define,
-        'import.meta.env.WAKU_ENTRIES_FILE': JSON.stringify(entriesFile),
-        'import.meta.env.WAKU_CONFIG_DIST_DIR': JSON.stringify(opts.distDir),
-        'import.meta.env.WAKU_CONFIG_PUBLIC_DIR': JSON.stringify(DIST_PUBLIC),
-      };
     },
     configResolved(config) {
+      entriesFile = `${config.root}/${opts.srcDir}/${SRC_ENTRIES}`;
       const { deploy, unstable_phase } = platformObject.buildOptions || {};
       if (
         (unstable_phase !== 'buildServerBundle' &&
           unstable_phase !== 'buildSsrBundle') ||
-        deploy !== 'cloudflare'
+        deploy !== 'deno'
       ) {
         return;
       }
@@ -76,6 +70,16 @@ export function deployDenoPlugin(opts: {
       config.ssr.resolve.conditions.push('worker');
       config.ssr.resolve.externalConditions ||= [];
       config.ssr.resolve.externalConditions.push('worker');
+    },
+    resolveId(source) {
+      if (source === `${opts.srcDir}/${SERVE_JS}`) {
+        return source;
+      }
+    },
+    load(id) {
+      if (id === `${opts.srcDir}/${SERVE_JS}`) {
+        return getServeJsContent(opts.distDir, DIST_PUBLIC, entriesFile);
+      }
     },
   };
 }
